requirements.txt
django==4.2.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==22.0.0